Create program-1.java
We use the integers , , and  to create the following series:

You are given  queries in the form of , , and . For each query, print the series corresponding to the given , , and  values as a single line of  space-separated integers.

Input Format

The first line contains an integer, , denoting the number of queries.
Each line  of the  subsequent lines contains three space-separated integers describing the respective , , and  values for that query.
 
Output Format

For each query, print the corresponding series on a new line. Each series must be printed in order as a single line of  space-separated integers.